Start switching from gflags to optparse
Re-write the nova.flags module to use optparse instead of gflags.

This provides an easier path to switching completely to optparse.
Next steps are to:

  1) Gradually switch each of the individual flags to optparse

  2) Re-use config code from other projects via openstack-common

optparse was chosen instead of argparse purely because that's what
the other projects use and that's what makes most sense for
openstack-common. Switching to argparse is something that can be
done later in openstack-common.

Change-Id: Ia49d42b4c7cc208fba140db6b8fd8f33c0f89e04