Fixes bug 888649
Change exception.VolumeIsBusy to derive from NovaException instead of
Error, so that an 'unexpected keyword' exception is not thrown when
the exception is raised with keyword parameters.

Add unit test to confirm that the 'unexpected keyword' exception is not
thrown when the exception is raised by nova.volume.driver.VolumeDriver

Responded to reviewer observations, fix pep8 errors in tset_volume.py,
added email address to Authors file.

Change-Id: I15464cb0cf72a2c71f430e4c8c5c2b27cd4e2ef9