vm_state:=error on driver exceptions during resize
Set vm_state to error if a driver raises an exception (of any type)
during resize_instance and finish_resize.

I want to highlight that there is a slight behavioral change here.
Before, on a MigrationError the exception was not reraised in the
resize_instance function. However, based on my understanding it is not
a big deal to make this change. The worst case is log duplication in
a fairly rare error state.

Change-Id: Ifd1ea66cf3f2e59ee06714f666e357a99936b4e3