Fix an error that prevents message from getting substituted.
The problem: self.__dict__ contains message.  This overrides the
default of None for the NovaException constructor.  Since a
message is provided, it's used directly instead of trying to
substitute the keyword arguments.  The solution is to limit the
keyword arguments we pass to the NovaException constructor, so
they don't contain message.  Fixes bug 917768.

Change-Id: I5986fa7b552c84ba7f902e78b32498ab294beb7b