Be more explicit about emptying connection pool.
This patch makes some code in test_qpid be more explicit about emptying
the connection pool.  It now ensures that it removes exactly how many
connections are in the pool.

The previous code made a couple of assumptions.  First, it assumed that
only one connection was in the pool (which is true, but it's still nice
not to make the assumption here in the cleanup code).  Second, it
assumed that free() returned the number of connections that have been
placed in the pool.  This is not correct.  The result also includes the
number of connections that could be created based on the max size of the
pool. Use the free_items attribute instead, which gives the exact number
of connections that have been put() in the pool.

Change-Id: I97378919c2d3e68f224862f07a75529575647163