Optionally pass a instance uuid to log methods.
This change is intended to make log messages including instance
uuids consistent. This eases operational support, as when a
customer asks for the history of there instance it is a simple
grep operation, instead of the current situation. I have included
a sample use, but will work through adding this to relevant log
messages over time.

Change-Id: I1f61c04f32dbb960471950ac7231313d9d9ced12