Handle --flagfile by converting to .ini style
Add code to translate flagfiles into .ini config files so that we can
deprecate and, eventually remove, --flagfile.

By using the flagfile compat code, we no longer need to process the
contents of flagfiles as CLI options and we can reduce the number
of CLI options to a sane set:

  --verbose
  --logdir
  --logfile
  --use_syslog
  --use_stderr
  --connection_type
  --sql_connection
  --api_paste_config
  --state_path
  --lock_path
  --fake_network
  --fake_rabbit

This, in turn, means we can remove the evil hacks which we needed in
order to register CLI options after the initial parsing of the command
line.

Change-Id: I9e24008fa634d7c8378b253c1f7a6d2169076086