Resolve bug/927714 -- get instance names from db.
I thought when I wrote the imagecache code for libvirt that the
instance names were of a fixed format. That is not correct. First
off the format is set in the instance_name_template flag. Worse
than that, it can change over time and old instances are not
renamed. So, now I check the database to see if the directory
name is a valid instance name (for this machine).

Change-Id: I318215aef5a10cee32959fd947ad628edff293a2