Fixes bug 949038
Make the api calls better fit the standard pattern where
read_deleted can be changed using context.read_deleted

I have retained the ability to pass read_deleted
explicitly. If that is not specified, it uses the value in
the context.

Note: read_deleted defaults to "no" in the context.

The two exceptions are:
 aggregate_host_get_all
 aggregate_get_all

In this case, it is better for read_deleted=yes to be the default
So in this case the context cannot be used, as that would default
to read_deleted=no. In this case you must explicity override
read_deleted, the context is totally ignored, as before.

Change-Id: Idb048a592d8c6b788651d131a3345e70989c0ec4