Stop using nova.exception from nova.rpc.
This patch is a part of continuing to remove dependencies from nova.rpc
on the rest of nova.  One RPC related exception was defined in
nova.exception, so that was moved to nova.rpc.common where the rest of
them live.  These exceptions were changed to no longer use NovaException
as their base.  Instead, there is a new RPCException base.

One other change that should be reviewed closely is the removal of using
nova.exception.wrap_exception() in nova.rpc.amqp.  As far as I can tell,
this didn't actually do anything since nothing was being passed in to
wrap_exception().

Change-Id: I36ff7c05ab0467ad8506b56d561c532eadf8dff8