Change base rpc version to 1.0 in compute rpcapi.
A recent patch change the RPC_API_VERSION in compute/rpcapi.py to 1.1.
This change was the right thing to do in compute/manager.py, but not
compute/rpcapi.py.  This patch changes it back to 1.0, and changes the
code to only specify that 1.1 is required for the new method.

This version in the rpcapi modules is used as the default version to set
in a message.  The version sent in a message is the minimum version
required to be implemented on the manager side for the message to be
successfully processed.  For most messages, that is just 1.0.  In the
case of the new method, it's 1.1.

This patch also changes the name of his variable in the rpcapi modules
from RPC_API_VERSION to BASE_RPC_API_VERSION to help clarify this a bit
more.

Change-Id: I2bb4602845869bd44b39531eeb4f1e7c0c4c20c2