Remove unused add_network_to_project() method
This method was added here:

  https://code.launchpad.net/~tr3buchet/nova/multi_nic/+merge/64767

but appears never to have been used.

We can safely remove it and the 'force' parameter to network_associate()
which only it uses.

This removal from the network service RPC API would not warrant a
major version number bump since nothing ever used this part of the
API.

Change-Id: Iba3cf7b61b629ba513a58a38886c9322818accb5