Add a link from HACKING to wiki GitCommitMessages page
The HACKING file contains a short example of an effective
commit message. For reasons for space, it cannot describe
the rationale behind this example. It also does not have
space to describe how to split up a patch into a series
of commits. Add a link from the HACKING file to the wiki

   http://wiki.openstack.org/GitCommitMessages

where further information can be found

Change-Id: I34d44485486b623b11743106f09d5ef631d35888