Reduce db access in prep_resize in the compute manager.
This patch changes the arguments passed to prep_resize in the compute
manager.  Previously it took instance and instance_type IDs.  It now
receives a full dict for both of these.  This cuts down on database
access needed on the compute node.

This method was not previously in the compute rpcapi module, so it has
been added there.  The scheduler had a bit of work done ot get it to use
the rpcapi module for prep_resize, as well.

Part of blueprint no-db-messaging.

Change-Id: Idadbf4fc624d5d1b128f758a46c61b3c840b9898