Re-work the handling of firewall_driver default
Fixes bug #1040430

We have a different default firewall driver for libvirt and
xenapi, yet the sample config file currently contains:

  firewall_driver=nova.virt.firewall.IptablesFirewallDriver

In the case of libvirt, it should actually be:

  firewall_driver=nova.virt.firewall.libvirt.IptablesFirewallDriver

This is really easy for users to get confused about.

Since we don't have a different sample config file for each hypervisor,
the best we can do is to just not include the default in the sample
config and have each hypervisor supply its default at runtime.

DocImpact: update nova.conf docs
Change-Id: Ie78371bcceac5a65978d695934e0246022f748a3