Fixes backwards compatible rpc schedule_run
Commit 8718f8e47d7d0504724495538eb320be3e209180 changed the
schedule_run_instance call to create all of the db records on the
compute host, but the old code path was left in for backwards
compatibility. Unfortunately an extra reservations parameter was
left in causing the backwards compatibility to be broken.

This patch takes out the reservations parameter and commits the
reservation in the backwards compatible call.

Fixes bug 1042302

Change-Id: I1e710d1e9a992f59a18eeedcd7d71726ff96c485