Allow VMs to be resumed after a hypervisor reboot
Fixes bug 1052696.

Update the compute manager to pass network_info and block_device_info
to the driver.resume() and update all virtualization drivers to accept
the new arguments.

For libvirt, change resume() to use _create_domain_and_network()
rather than _create_domain(). This eliminates the assumption that the
network and block device connections remained in place from the period
between the VM being suspended and resumed.  Instead, all the
networking and block connections will be rebuilt on resume (in case
they are missing) as is the case after a hypervisor reboot.

Change-Id: I6e19ec42f7e929678abce8f276c0a6e91f1fa8af