Add scope to extra_specs entries
Do to conflicts between different scheduler filters it is necessary
to allow an optional scope for the different keys.  This scope is a
leading string followed by a ':' character, e.g.:

        foo:bar

is the `bar' entry with scope `foo'.

The Trusted filter will now use the scope `trust' and the
AggregateInstanceExtraSpecs and ComputeCapabilities filters will
check against any unscoped keys in the `extra_specs' table.
Any new filters that utilize the `extra_specs' table will need to
use a unique scope string for any keys they require.

Resolves bug 1039386

Change-Id: I2466dc3d4de8e9aeb76b294eeda1c939c0413366
Signed-off-by: Don Dugger <donald.d.dugger@intel.com