Allows waiting timers in libvirt to raise NotFound
There are cases where an operation will fail when communicating with
libvirt. We were eating the exception even though the operation
failed, which has the potential to put the instance into an
unrecoverable state.

This patch allows NotFound exceptions to propogate up so that they
are caught by the state handling code and the task state can be
set to error.

Fixes bug 1002814

Change-Id: Iddc319b24aee0b7132155f50b9d3b0eee9bb3fa8