Add scheduler retries for prep_resize operations.
Retry operations are subject to race conditions for compute resources.
prep_resize can race for resources on the destination host and
revert_resize can race for resources on the original source host.
This type of problem is handled for new instance build requests
by the resource tracker in compute.

As a first step to extending the resource tracker to cover resizes,
this patch adds scheduler retries around the prep_resize operation.

bug 1065267

Change-Id: I1b32c0a282772d9580e322b50990932016058329